Replaces all previous scripts by a few new ones
The numerous scripts have been replaced by only 3 scripts:
* tizen-info
* tizen-info-gerrit-status
* tizen-info-multiuser

tizen-info is the main script which replaces most of the previous ones.
It allows the generation of a status of all commits matching a given
request.

tizen-info-gerrit-status generates a status of the pending reviews. It
uses tizen-info.

tizen-info-multiuser generates a status of the multi-user commits.

Note that the previous scripts are still available on branch
v1_2014-03-25.

Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>